refactor(utkast): flytt til vertikal slice-arkitektur
Flytt fra lagdelt struktur til vertikale slices: feature-kode ligger nå i src/utkast/ (co-lokert komponent, typer, tekster, fetch, utils og CSS med ingen/, fallback/ og link-card/ som undervarianter), og delt infrastruktur i src/shared/. i18n-tekster er splittet i feature- og shell-strenger, og fetch-logikken er skilt ut i utkastFetch.ts. Alle imports og e2e-referanser er oppdatert.
