--- src/assets/KattIBoks.tsx
@@ -1,211 +1,0 @@
-import { useId } from "react";
-
-interface Props {
-  alt: string;
-}
-
-const KattIBoks = ({ alt }: Props) => (
-  <svg
-    width="245"
-    height="299"
-    viewBox="0 0 245 299"
-    fill="none"
-    aria-labelledby="katt-i-boks"
-    role="img"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <title id={useId()}>{alt}</title>
-    <path
-      d="M167.687 196.258C167.687 221.347 167.687 263.229 122.259 263.229C76.8311 263.229 76.8311 221.347 76.8311 196.258C76.8311 171.169 97.1699 150.83 122.259 150.83C147.348 150.83 167.687 171.169 167.687 196.258Z"
-      fill="#262626"
-    />
-    <path
-      d="M122.259 151.064C90.0379 151.064 39.3645 180.881 18.0554 195.79L29.0612 214.055C96.2667 225.295 119.215 195.79 119.215 195.79C119.215 195.79 124.367 167.3 122.259 151.064Z"
-      fill="#262626"
-    />
-    <circle
-      r="12.6449"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 18.4212 207.989)"
-      fill="#262626"
-    />
-    <circle
-      r="5.15164"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 7.03746 209.852)"
-      fill="#262626"
-    />
-    <circle
-      r="3.74665"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 7.76266 201.705)"
-      fill="#262626"
-    />
-    <circle
-      r="2.34166"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 7.76277 201.705)"
-      fill="#FFB8B8"
-    />
-    <circle
-      r="3.74665"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 18.2375 218.911)"
-      fill="#262626"
-    />
-    <circle
-      r="5.15164"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 10.7838 216.62)"
-      fill="#262626"
-    />
-    <circle
-      r="7.10248"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 18.6252 207.365)"
-      fill="#FFB8B8"
-    />
-    <circle
-      r="3.05072"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 6.86969 209.792)"
-      fill="#FFB8B8"
-    />
-    <circle
-      r="3.27987"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 10.705 216.618)"
-      fill="#FFB8B8"
-    />
-    <circle
-      r="2.34166"
-      transform="matrix(-0.866025 0.5 0.5 0.866025 18.2374 218.911)"
-      fill="#FFB8B8"
-    />
-    <path
-      d="M122.741 151.064C154.962 151.064 205.636 180.881 226.945 195.79L215.939 214.055C148.734 225.295 125.785 195.79 125.785 195.79C125.785 195.79 120.634 167.3 122.741 151.064Z"
-      fill="#262626"
-    />
-    <circle
-      cx="226.579"
-      cy="207.989"
-      r="12.6449"
-      transform="rotate(30 226.579 207.989)"
-      fill="#262626"
-    />
-    <circle
-      cx="237.963"
-      cy="209.852"
-      r="5.15164"
-      transform="rotate(30 237.963 209.852)"
-      fill="#262626"
-    />
-    <circle
-      cx="237.238"
-      cy="201.705"
-      r="3.74665"
-      transform="rotate(30 237.238 201.705)"
-      fill="#262626"
-    />
-    <circle
-      cx="237.237"
-      cy="201.705"
-      r="2.34166"
-      transform="rotate(30 237.237 201.705)"
-      fill="#FFB8B8"
-    />
-    <circle
-      cx="226.762"
-      cy="218.911"
-      r="3.74665"
-      transform="rotate(30 226.762 218.911)"
-      fill="#262626"
-    />
-    <circle
-      cx="234.216"
-      cy="216.62"
-      r="5.15164"
-      transform="rotate(30 234.216 216.62)"
-      fill="#262626"
-    />
-    <circle
-      cx="226.375"
-      cy="207.365"
-      r="7.10248"
-      transform="rotate(30 226.375 207.365)"
-      fill="#FFB8B8"
-    />
-    <circle
-      cx="238.13"
-      cy="209.792"
-      r="3.05072"
-      transform="rotate(30 238.13 209.792)"
-      fill="#FFB8B8"
-    />
-    <circle
-      cx="234.295"
-      cy="216.618"
-      r="3.27987"
-      transform="rotate(30 234.295 216.618)"
-      fill="#FFB8B8"
-    />
-    <circle
-      cx="226.763"
-      cy="218.911"
-      r="2.34166"
-      transform="rotate(30 226.763 218.911)"
-      fill="#FFB8B8"
-    />
-    <path
-      d="M128 161.861L131.224 182.687C131.511 184.539 130.803 189.15 127.1 189.724C122.472 190.44 121.401 186.578 121.042 184.264L117.819 163.437C111.514 122.709 134.201 102.223 160.592 96.8662C185.173 91.8763 185.219 73.5178 185.219 67.8978C185.219 54.5937 175.305 45.4309 162.226 44.1434C149.147 42.8559 139.478 36.4475 139.478 25.5269C139.478 13.5845 149.5 6.65507 160.592 6.65506C185.219 6.65503 207.5 23.5736 210.196 57.1392C212.891 90.7047 188.416 108.943 172.261 112.871C161.528 115.48 122.842 128.538 128 161.861Z"
-      fill="#262626"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M74.2304 216.447L36.5298 241.654L60.1805 240.332L74.3397 230.24L74.2304 216.447Z"
-      fill="#A86400"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M170.707 216.447L208.282 241.654L185.225 240.098L170.473 230.24L170.707 216.447Z"
-      fill="#A86400"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M170.671 298.074H74.0101L68.6106 261.833L74.0101 216.396L107.482 219.023H137.455L170.671 216.396L176.093 285.142L170.671 298.074Z"
-      fill="#D47B00"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M146.525 286.936H150.246V275.798H146.525V286.936Z"
-      fill="#523300"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M157.688 286.936H161.408V275.798H157.688V286.936Z"
-      fill="#523300"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M144.665 294.361H163.269V290.648H144.665V294.361Z"
-      fill="#523300"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M148.5 271.074L145 277.074H152L148.5 271.074Z"
-      fill="#523300"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M159.5 271.074L156 277.074H163L159.5 271.074Z"
-      fill="#523300"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M130.886 238.672H113.795C112.351 238.672 111.178 239.918 111.178 241.457C111.178 242.995 112.351 244.241 113.795 244.241H130.886C132.332 244.241 133.503 242.995 133.503 241.457C133.503 239.918 132.332 238.672 130.886 238.672"
-      fill="#262626"
-    />
-  </svg>
-);
-
-export default KattIBoks;
