--- src/language/text.ts
@@ -1,107 +1,0 @@
-export const text = {
-  utkast: {
-    nb: "Utkast",
-    en: "Drafts",
-    nn: "Utkast",
-  },
-  description: {
-    nb: "Her kan du fortsette på skjemaer du ikke har sendt til oss ennå",
-    en: "Here you can continue with the forms you have not yet submitted to us",
-    nn: "Her kan du halde fram med skjema du ikkje har sendt inn til oss enno",
-  },
-  started: {
-    nb: "Påbegynt",
-    en: "Started",
-    nn: "Starta på",
-  },
-  emptyKitten: {
-    nb: "En svart katt som gjemmer seg bak ett papirark",
-    en: "A black cat hiding behind a piece of paper",
-    nn: "En svart katt som gjemmer seg bak ett papirark",
-  },
-  emptySubheading: {
-    nb: "Ikke noe i utkast",
-    en: "Nothing in drafts",
-    nn: "Ikkje noko i utkast",
-  },
-  emptyOtherServices: {
-    nb: "Pengestøtte og tjenester A-Å",
-    en: "All services and benefits",
-    nn: "Pengestøtte og tenester A-Å",
-  },
-  loading: {
-    nb: "Henter innhold ...",
-    en: "Fetching content ...",
-    nn: "Hentar innhald ...",
-  },
-  errorHeading: {
-    nb: "Noe gikk galt!",
-    en: "Something went wrong!",
-    nn: "Noko gjekk gale",
-  },
-  errorText: {
-    nb: "Vi har for øyeblikket tekniske problemer. Dette kan føre til at du ikke får opp all informasjon. Vennligst prøv igjen senere.",
-    en: "We're having technical difficulties. Some information may not be available, please try again later.",
-    nn: "Vi har for øyeblikket tekniske problemer. Dette kan føre til at du ikke får opp all informasjon. Vennligst prøv igjen senere.",
-  },
-  errorHelp: {
-    nb: "Prøv igjen senere eller gå til ",
-    en: "Try again later or go to ",
-    nn: "Prøv igjen seinare eller gå til ",
-  },
-  errorKitten: {
-    nb: "En svart katt som har forsøkt å hoppe inn i en boks med hodet først men satt seg fast slik at bare føttene og halen stikker opp",
-    en: "A black cat that has attempted to jump into a box head first but the box was too smal so that her feet ant tail is up in the air",
-    nn: "En svart katt som har forsøkt å hoppe inn i en boks med hodet først men satt seg fast slik at bare føttene og halen stikker opp",
-  },
-  minSide: {
-    nb: "Min side",
-    en: "My Page",
-    nn: "Mi side",
-  },
-  opprettet: {
-    nb: "Du startet på utkastet ",
-    en: "You started the draft ",
-    nn: "Du starta på utkastet ",
-  },
-  slettes: {
-    nb: "Utkastet slettes ",
-    en: "The draft will be deleted ",
-    nn: "Utkastet slettes ",
-  },
-  endret: {
-    nb: "Endret ",
-    en: "Edited ",
-    nn: "Endra ",
-  },
-  ikkeEndret: {
-    nb: "Ikke endret siden opprettelse",
-    en: "Not edited after creation",
-    nn: "Itje endra sia oppretting",
-  },
-  ingenUtkastTittel: {
-    nb: "Ingen utkast",
-    en: "No drafts",
-    nn: "Ingen utkast",
-  },
-  ingenUtkastIngress: {
-    nb: "Vi kan ikke se at du har noen påbegynte søknader",
-    en: "We can not find any started drafts",
-    nn: "Vi kan itje sjå at du har nokon søknader du har byrja på",
-  },
-  ingenUtkastLenketekst: {
-    nb: "Gå til søknad og skjema",
-    en: "Go to applications and forms",
-    nn: "Gå til søknad og skjema",
-  },
-  partialContent: {
-    nb: "Noen av systemene våre har problemer for øyeblikket. ",
-    en: "Some of our systems are currently experiencing issues.",
-    nn: "Noen av systemene våre har problemer for øyeblikket.",
-  },
-  pencil: {
-    nb: "Blyant",
-    en: "Pencil",
-    nn: "Blyant",
-  },
-};
