+++ src/pages/[locale]/_error.module.css
