+++ src/pages/[locale]/_index.module.css
