+++ src/pages/error.astro
@@ -1,11 +1,11 @@
 ---
 import { MIN_SIDE_URL } from "astro:env/server";
 import { BodyLong, Heading } from "@navikt/ds-react";
-import KattIBoks from "../assets/KattIBoks";
-import type { Language } from "../language/language";
-import { text } from "../language/text";
+import KattIBoks from "@src/shared/assets/KattIBoks";
+import type { Language } from "@src/shared/language/language";
+import { text } from "@src/shared/language/text";
 import Layout from "../layouts/Layout.astro";
-import styles from "../styles/error.module.css";
+import styles from "./[locale]/_error.module.css";
 
 const language = Astro.currentLocale as Language;
 ---
