+++ src/pages/index.astro
@@ -1,3 +1,1 @@
----
 
----
