+++ src/shared/assets/Pencil.tsx
@@ -5,13 +5,7 @@ interface Props {
 }
 
 const Pencil = ({ alt }: Props) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="20"
-    height="20"
-    viewBox="0 0 24 24"
-    fill="none"
-  >
+  <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none">
     <title id={useId()}>{alt}</title>
     <rect width="20" height="20" rx="2" fill="#005B82" />
     <path
