+++ src/shared/language/text.ts
@@ -1,0 +1,37 @@
+export const text = {
+  utkast: {
+    nb: "Utkast",
+    en: "Drafts",
+    nn: "Utkast",
+  },
+  description: {
+    nb: "Her kan du fortsette på skjemaer du ikke har sendt til oss ennå",
+    en: "Here you can continue with the forms you have not yet submitted to us",
+    nn: "Her kan du halde fram med skjema du ikkje har sendt inn til oss enno",
+  },
+  minSide: {
+    nb: "Min side",
+    en: "My Page",
+    nn: "Mi side",
+  },
+  errorHeading: {
+    nb: "Noe gikk galt!",
+    en: "Something went wrong!",
+    nn: "Noko gjekk gale",
+  },
+  errorText: {
+    nb: "Vi har for øyeblikket tekniske problemer. Dette kan føre til at du ikke får opp all informasjon. Vennligst prøv igjen senere.",
+    en: "We're having technical difficulties. Some information may not be available, please try again later.",
+    nn: "Vi har for øyeblikket tekniske problemer. Dette kan føre til at du ikke får opp all informasjon. Vennligst prøv igjen senere.",
+  },
+  errorHelp: {
+    nb: "Prøv igjen senere eller gå til ",
+    en: "Try again later or go to ",
+    nn: "Prøv igjen seinare eller gå til ",
+  },
+  errorKitten: {
+    nb: "En svart katt som har forsøkt å hoppe inn i en boks med hodet først men satt seg fast slik at bare føttene og halen stikker opp",
+    en: "A black cat that has attempted to jump into a box head first but the box was too smal so that her feet ant tail is up in the air",
+    nn: "En svart katt som har forsøkt å hoppe inn i en boks med hodet først men satt seg fast slik at bare føttene og halen stikker opp",
+  },
+};
