+++ src/shared/utils/analytics.ts
@@ -3,9 +3,7 @@ import { getAnalyticsInstance } from "@navikt/nav-dekoratoren-moduler";
 
 type ExtendedAmpltitudeEvent = AmplitudeEvent<"navigere", { kategori: string }>;
 
-const analyticsLogger = getAnalyticsInstance<ExtendedAmpltitudeEvent>(
-  "tms-utkast-frontend",
-);
+const analyticsLogger = getAnalyticsInstance<ExtendedAmpltitudeEvent>("tms-utkast-frontend");
 
 export const logEvent = async (metric: string) => {
   await analyticsLogger("navigere", { kategori: metric });
