+++ src/shared/utils/environment.ts
