+++ src/utkast/fallback/Error.astro
@@ -1,7 +1,7 @@
 ---
 import { Alert, BodyLong } from "@navikt/ds-react";
-import { type Language } from "@src/language/language";
-import { text } from "@src/language/text.ts";
+import { type Language } from "../../shared/language/language";
+import { text } from "../../shared/language/text";
 import styles from "./Error.module.css";
 
 interface Props {
