+++ src/utkast/ingen/IngenUtkast.astro
@@ -1,8 +1,8 @@
 ---
 import { BodyLong, Heading } from "@navikt/ds-react";
-import TomtKatt from "@src/assets/TomtKatt";
-import type { Language } from "@src/language/language.ts";
-import { text } from "@src/language/text.ts";
+import TomtKatt from "../../shared/assets/TomtKatt";
+import type { Language } from "../../shared/language/language";
+import { text } from "../utkastText";
 import styles from "./IngenUtkast.module.css";
 
 const language = Astro.currentLocale as Language;
