+++ src/utkast/link-card/UtkastLinkCard.tsx
@@ -1,28 +1,21 @@
 import { LinkCard, Tag } from "@navikt/ds-react";
-import type { Language } from "@src/language/language";
-import { text } from "@src/language/text";
-import { logEvent } from "@src/utils/analytics";
 import dayjs from "dayjs";
-import styles from "./Utkast.module.css";
-import type { UtkastElement } from "./UtkastTypes";
+import type { Language } from "../../shared/language/language";
+import { logEvent } from "../../shared/utils/analytics";
+import styles from "../Utkast.module.css";
+import { text } from "../utkastText";
+import type { UtkastElement } from "../utkastTypes";
 
 interface UtkastLinkCardProps {
   utkast: UtkastElement;
   language: Language;
 }
 
-export default function UtkastLinkCard({
-  utkast,
-  language,
-}: UtkastLinkCardProps) {
+export default function UtkastLinkCard({ utkast, language }: UtkastLinkCardProps) {
   const dateFormatter = (date: string) => dayjs(date).format("DD.MM.YYYY");
 
   return (
-    <LinkCard
-      onClick={() => logEvent("utkast-åpnet")}
-      data-color="accent"
-      className={styles.linkCard}
-    >
+    <LinkCard onClick={() => logEvent("utkast-åpnet")} data-color="accent" className={styles.linkCard}>
       <LinkCard.Title>
         <LinkCard.Anchor href={utkast.link}>{utkast.tittel}</LinkCard.Anchor>
       </LinkCard.Title>
