+++ src/utkast/utkastText.ts
@@ -1,0 +1,32 @@
+export const text = {
+  opprettet: {
+    nb: "Du startet på utkastet ",
+    en: "You started the draft ",
+    nn: "Du starta på utkastet ",
+  },
+  slettes: {
+    nb: "Utkastet slettes ",
+    en: "The draft will be deleted ",
+    nn: "Utkastet slettes ",
+  },
+  emptyKitten: {
+    nb: "En svart katt som gjemmer seg bak ett papirark",
+    en: "A black cat hiding behind a piece of paper",
+    nn: "En svart katt som gjemmer seg bak ett papirark",
+  },
+  ingenUtkastTittel: {
+    nb: "Ingen utkast",
+    en: "No drafts",
+    nn: "Ingen utkast",
+  },
+  ingenUtkastIngress: {
+    nb: "Vi kan ikke se at du har noen påbegynte søknader",
+    en: "We can not find any started drafts",
+    nn: "Vi kan itje sjå at du har nokon søknader du har byrja på",
+  },
+  ingenUtkastLenketekst: {
+    nb: "Gå til søknad og skjema",
+    en: "Go to applications and forms",
+    nn: "Gå til søknad og skjema",
+  },
+};
