+++ src/utkast/utkastUtils.ts
@@ -1,12 +1,5 @@
 import dayjs from "dayjs";
-
-export interface UtkastElement {
-  tittel: string;
-  link: string;
-  utkastId: string;
-  opprettet: string;
-  sistEndret: string;
-}
+import type { UtkastElement } from "./utkastTypes";
 
 export const sortByOpprettet = (a: UtkastElement, b: UtkastElement) =>
   dayjs(a.opprettet).isAfter(dayjs(b.opprettet)) ? -1 : 1;
